testing-service/requirements.txt
requests
python-dotenv
pydantic-settings
pytest